pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "modx"
version = "0.1.0"
description = "Autonomous Codebase Modernizer"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "click>=8.0.0",
    "colorama>=0.4.0",
    "pyyaml>=6.0",
    "gitpython>=3.1.0",
    "requests>=2.25.0",
    "rich>=10.0.0",
]
authors = [
    {name = "ModX Team"},
]
maintainers = [
    {name = "ModX Team"},
]
keywords = ["cli", "modernization", "refactoring", "codebase"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[project.scripts]
modx = "modx.cli:cli"

[tool.setuptools.packages.find]
where = ["."]